Add files via upload
Morse code is a code where each letter of the English alphabet, each digit, and various
punctuation characters are represented by a series of dots and dashes. Table 9-14 shows
part of the code. Write a program that asks the user to enter a string, and then converts that
string to Morse code. Use hyphens for dashes and periods for dots. Please look up Morse code table if you want to